Refactor SignMessageLib contract
Move encode, estimateGas + getAddress functions into abstract parent classes.
