+++ packages/protocol-kit/src/adapters/web3/contracts/BaseContractWeb3.ts
@@ -7,9 +7,11 @@ import Web3Adapter from '@safe-global/protocol-kit/adapters/web3/Web3Adapter'
 import { SafeVersion } from '@safe-global/safe-core-sdk-types'
 import {
   EncodeFunction,
+  EstimateGasFunction,
   GetAddressFunction
 } from '@safe-global/protocol-kit/contracts/AbiType/common/BaseContract'
 import BaseContract from '@safe-global/protocol-kit/adapters/BaseContract'
+import { Web3TransactionOptions } from '@safe-global/protocol-kit/adapters/web3/types'
 
 /**
  * Abstract class BaseContractWeb3 extends BaseContract to specifically integrate with the Web3.js library.
@@ -66,6 +68,16 @@ abstract class BaseContractWeb3<
 
   encode: EncodeFunction<ContractAbiType> = (functionToEncode, args) => {
     return this.contract.methods[functionToEncode](...(args as Array<any>)).encodeABI()
+  }
+
+  estimateGas: EstimateGasFunction<ContractAbiType, Web3TransactionOptions> = (
+    functionToEstimate,
+    args,
+    options = {}
+  ) => {
+    return this.contract.methods[functionToEstimate](...(args as Array<any>))
+      .estimateGas(options)
+      .then(BigInt)
   }
 }
