+++ packages/protocol-kit/src/adapters/web3/contracts/SignMessageLib/SignMessageLibBaseContractWeb3.ts
@@ -1,30 +1,28 @@
-import Contract from 'web3-eth-contract'
+import { Abi } from 'abitype'
 import { AbiItem } from 'web3-utils'
 
 import Web3Adapter from '@safe-global/protocol-kit/adapters/web3/Web3Adapter'
 import { SafeVersion } from '@safe-global/safe-core-sdk-types'
-import SignMessageLibBaseContract from '@safe-global/protocol-kit/adapters/SignMessageLibBaseContract'
+import BaseContractWeb3 from '@safe-global/protocol-kit/adapters/web3/contracts/BaseContractWeb3'
+import { contractName } from '@safe-global/protocol-kit/contracts/config'
 
 /**
- * Abstract class SignMessageLibBaseContractWeb3 extends SignMessageLibBaseContract to specifically integrate with the Web3.js v6 library.
+ * Abstract class SignMessageLibBaseContractWeb3 extends BaseContractWeb3 to specifically integrate with the SignMessageLib contract.
  * It is designed to be instantiated for different versions of the SignMessageLib contract.
- *
- * This abstract class sets up the Web3 v6 Contract object that interacts with a SignMessageLib contract version.
  *
  * Subclasses of SignMessageLibBaseContractWeb3 are expected to represent specific versions of the SignMessageLib contract.
  *
  * @template SignMessageLibContractAbiType - The ABI type specific to the version of the SignMessageLib contract, extending InterfaceAbi from Web3.
- * @extends SignMessageLibBaseContract<SignMessageLibContractAbiType> - Extends the generic SignMessageLibBaseContract with Web3-specific implementation.
+ * @extends BaseContractWeb3<SignMessageLibContractAbiType> - Extends the generic BaseContractWeb3.
  *
  * Example subclasses:
  * - SignMessageLibContract_v1_4_1_Web3 extends SignMessageLibBaseContractWeb3<SignMessageLibContract_v1_4_1_Abi>
  * - SignMessageLibContract_v1_3_0_Web3 extends SignMessageLibBaseContractWeb3<SignMessageLibContract_v1_3_0_Abi>
  */
 abstract class SignMessageLibBaseContractWeb3<
-  SignMessageLibContractAbiType extends AbiItem[]
-> extends SignMessageLibBaseContract<SignMessageLibContractAbiType> {
-  contract: Contract
-  adapter: Web3Adapter
+  SignMessageLibContractAbiType extends AbiItem[] & Abi
+> extends BaseContractWeb3<SignMessageLibContractAbiType> {
+  contractName: contractName
 
   /**
    * @constructor
@@ -45,10 +43,19 @@ abstract class SignMessageLibBaseContractWeb3<
     customContractAddress?: string,
     customContractAbi?: SignMessageLibContractAbiType
   ) {
-    super(chainId, defaultAbi, safeVersion, customContractAddress, customContractAbi)
+    const contractName = 'signMessageLibVersion'
 
-    this.adapter = web3Adapter
-    this.contract = web3Adapter.getContract(this.contractAddress, this.contractAbi)
+    super(
+      contractName,
+      chainId,
+      web3Adapter,
+      defaultAbi,
+      safeVersion,
+      customContractAddress,
+      customContractAbi
+    )
+
+    this.contractName = contractName
   }
 }
